arreglo de bugs en ClientesManager::modificar y cambio de int por long long en cuil

+++ farmaceutica/include/ArchivoClientes.h
@@ -23,7 +23,7 @@ public:
     bool modificarCliente(Cliente, int pos);
 
 
-    int buscarPorCUIL(int cuilSolicitado);
+    int buscarPorCUIL(long long cuilSolicitado);
 
     bool existe() const;
 };
